refactor(showLaps): create a 'createLapNode' helper function
